% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{make_subplot}
\alias{make_subplot}
\title{Helper function for making subplots}
\usage{
make_subplot(var, prh)
}
\arguments{
\item{var}{Variable for subplot}

\item{prh}{PRH object}
}
\value{
A plotly object
}
\description{
Helper function for making subplots
}
