% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mn200312_58}
\alias{mn200312_58}
\title{Sample PRH data.}
\format{
A tibble with columns:
\itemize{
\item dn (num) Matlab datenum of record
\item p (num) pressure, or approximate depth in meters
\item aw (num x 3) a 3 column matrix with x, y, and z accelerometry in whale frame
\item mw (num x 3) a 3 column matrix with x, y, and z magnetometry in whale frame
\item gw (num x 3) a 3 column matrix with x, y, and z gyroscope in whale frame
\item speed (num) speed of the whale in m/s (as estimated by the jiggle method)
\item pitch (num) pitch of the whale in radians
\item roll (num) roll of the whale in radians
\item head (num) heading of the whale in radians
\item dt (POSIXct) datetime of record
\item secs (numeric) seconds since start of deployment
\item jerk (num) norm of the jerk vector
And attributes:
\item whaleid (chr) e.g. "mn200312-58"
\item fs (num) sampling rate in Hz
\item tz (chr) local timezone e.g. "Etc/GMT+3" for UTC-3
}

See \code{\link{read_nc}} for more details.
}
\usage{
mn200312_58
}
\description{
The PRH for a humpback whale tagged in Antarctica on March 12, 2020.
}
\keyword{datasets}
