% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_nc}
\alias{read_nc}
\title{Read a CATS PRH from a NetCDF file}
\usage{
read_nc(nc_path)
}
\arguments{
\item{nc_path}{path to the CATS PRH}
}
\value{
A tibble with columns:
\itemize{
\item dn (num) Matlab datenum of record
\item p (num) pressure, or approximate depth in meters
\item aw (num x 3) a 3 column matrix with x, y, and z accelerometry in whale frame
\item mw (num x 3) a 3 column matrix with x, y, and z magnetometry in whale frame
\item gw (num x 3) a 3 column matrix with x, y, and z gyroscope in whale frame
\item speed (num) speed of the whale in m/s (as estimated by the jiggle method)
\item pitch (num) pitch of the whale in radians
\item roll (num) roll of the whale in radians
\item head (num) heading of the whale in radians
\item dt (POSIXct) datetime of record
\item secs (numeric) seconds since start of deployment
\item jerk (num) norm of the jerk vector
\item x (num) x coordinate of dead-reckoned track, relative to deployment location (m; eastings)
\item y (num) y coordinate of dead-reckoned track, relative to deployment location (m; northings)
\item z (num) z coordinate of dead-reckoned track, relative to deployment location (m; depth)
}

And attributes:
\itemize{
\item whaleid (chr) e.g. "mn200312-58"
\item fs (num) sampling rate in Hz
\item tz (chr) local timezone e.g. "Etc/GMT+3" for UTC-3
}
}
\description{
Read a CATS PRH from a NetCDF file
}
\examples{
\dontrun{
nc_path <- system.file("extdata", "mn200312-58_prh10.nc", package = "catsr")
mn200312_58 <- read_nc(nc_path)
}
}
