% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{view_cats}
\alias{view_cats}
\title{Visualize CATS data}
\usage{
view_cats(prh, vars = "p")
}
\arguments{
\item{prh}{A CATS PRH object (see \code{\link{read_nc}})}

\item{vars}{A vector of columns in \code{prh} to plot. "p" by default (depth).}
}
\value{
A plotly figure, with subplots in one column. One PRH variable per
subplot. Triaxial variables (e.g. \code{aw}) represented as multiple colors in one
subplot.
}
\description{
Visualize CATS data
}
\examples{
view_cats(mn200312_58, c("p", "pitch", "aw"))
}
