% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{view_cats_3d}
\alias{view_cats_3d}
\title{Visualize CATS data as 3d geotrack}
\usage{
view_cats_3d(prh)
}
\arguments{
\item{prh}{A CATS PRH object (see \code{\link{read_nc}})}
}
\value{
A 3d plotly figure
}
\description{
Visualize CATS data as 3d geotrack
}
\examples{
view_cats_3d(mn200312_58)
}
