% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{view_cats_3d}
\alias{view_cats_3d}
\title{Plot a georeferenced track in 3d}
\usage{
view_cats_3d(prh, width = 1)
}
\arguments{
\item{prh}{A CATS PRH object}

\item{width}{Width of the track (in meters)}
}
\value{
A plotly 3d mesh figure
}
\description{
Plot a georeferenced track in 3d
}
\examples{
view_cats_3d(mn200312_58, width = 10)
}
